+++ src/App.tsx
@@ -38,6 +38,7 @@ import HqRequestsLeader from './pages/hq/RequestsLeader'
 import HqRequestsPartnerByLeader from './pages/hq/RequestsPartnerByLeader'
 import HqRequestsPartnerDirect from './pages/hq/RequestsPartnerDirect'
 import HqRequestsMerchantDirect from './pages/hq/RequestsMerchantDirect'
+import HqRequestResultLog from './pages/hq/RequestResultLog'
 import HqLeaders from './pages/hq/Leaders'
 import HqPartners from './pages/hq/Partners'
 import HqMerchants from './pages/hq/Merchants'
@@ -66,6 +67,7 @@ const HQ_PAGES: Record<string, JSX.Element> = {
   'requests/partner-by-leader': <HqRequestsPartnerByLeader />,
   'requests/partner-direct': <HqRequestsPartnerDirect />,
   'requests/merchant-direct': <HqRequestsMerchantDirect />,
+  'requests/result-log': <HqRequestResultLog />,
   leaders: <HqLeaders />,
   'leaders/sales': <HqLeaderSales />,
   partners: <HqPartners />,